fixed bugs

+++ main/GameLogic/CardClass.cs
@@ -6,7 +6,7 @@ namespace CardClass{
     public class Card{
         public string? Id { get; set; }
         public string? Name { get; set; }
-        public float Damage { get; set; }
+        public double Damage { get; set; }
         public ElementType? Element { get; set; }
         public string? Type { get; set; }
     }
